Se limpio y se comento el codigo de automatizacion del proyecto sample de lippia que no va a ser usado.
