+++ src/main/java/lippia/web/constants/BookstoreConstants.java
@@ -1,0 +1,9 @@
+package lippia.web.constants;
+
+public class BookstoreConstants {
+
+    /*public static final String SEARCH_BUTTON_NAME = "name:btnK";
+    public static final String STATS_ID = "id:rcnt";
+    /*
+
+}
