--- src/main/java/lippia/web/constants/GoogleConstants.java
@@ -1,10 +1,0 @@
-package lippia.web.constants;
-
-public class GoogleConstants {
-
-    public static final String INPUT_SEARCH_XPATH = "xpath://textarea[@class='gLFyf']";
-    public static final String SEARCH_BUTTON_NAME = "name:btnK";
-    public static final String STATS_ID = "id:rcnt";
-
-
-}
