+++ src/main/java/lippia/web/services/BookstoreLogin.java
@@ -1,0 +1,10 @@
+package lippia.web.services;
+
+import com.crowdar.core.actions.ActionManager;
+import lippia.web.constants.BookstoreConstants;
+import org.openqa.selenium.WebElement;
+
+public class BookstoreLogin extends ActionManager {
+
+
+}
