+++ src/main/java/lippia/web/services/BookstoreMyAccountService.java
@@ -1,0 +1,10 @@
+package lippia.web.services;
+
+import com.crowdar.core.actions.ActionManager;
+import lippia.web.constants.BookstoreConstants;
+import org.openqa.selenium.WebElement;
+import org.testng.Assert;
+
+public class BookstoreMyAccountService extends ActionManager {
+
+}
